requirements.txt
crewai==0.28.8
streamlit==1.29.0
python-dotenv==1.0.0
langchain-google-genai==0.0.11
google-generativeai==0.3.2
streamlit-option-menu==0.3.6
plotly==5.18.0
pandas==2.1.4
langchain==0.1.7
python-decouple==3.8
pydantic==2.5.3